Fix issue where some bytes were not being read
Bug: When calculating the total loops needed to read the byte, I used
the total chunk size / buffer size value, which will ignore the
remainder bytes if chunk size was not a multiple of the buffer size.

Fix: This change adds an ugly fix for this problem by checking whether
there are any extra bytes remaining, and alloting another slot of buffer
states array to store that value.