Make it work on binary files
So I had a hunch that may be this is related with opening the file with
a particular mode (rb vs r), but there is no such thing for the `open`
syscall which is what Go calls under the hood when using `os.Open`.

I recalled from Rob Pike's Go Proverbs talk where he mentions that
there's a unicode isspace function that is copied in some other package
instead of importing the unicode package. So I took a look at the source
code, and there it was. I was missing out a couple of whitespace
characters. Adding this ensures the word count for binary files is
correct!

The two whitespace characters are 0xA0 (nbsp), and 0x85 (next line).