Fix the race condition properly
The actual problem was that the `wg.Done()` was getting called inside
the function. But this does not guarantee that the function return
happens immediately. OS scheduling can mean that the go rountine threads
might be in the "returning" state, and because all the go routines have
already called `wg.Done()`, the blocking ends, and the for loop might
start. This is the reason for race condition. So the correct fix would
be to call `wg.Done()` right after the `processBuffer` function
returned.