Simplify iteration logic
For each buffer counting routine, for the first byte in the thread, if
we can set the first and last bytes to the same value, and update last
byte for every step, we can get rid of the 'isPreviousCharSpace' flag
entirely. There is a cost to finding if a last char is a space by
calling the function the byte directly, but since this happens only when
the current char is a space, it shouldn't be too bad.