Fix issue with off-by-one error in word count in binary files
In the reduce step, we are not checking the last character of the last
chunk to see if there is a space. If it's not a space, we should
increment the value. This showed up only in binary files. Presumably
because it doesn't have an ending "\n" which is getting added in text
files created using my editor